fix: bug on userpage
